Create Message.java
This takes in the array of messages and lowercases them and takes out any non-alphabetical characters. I took some of the string messages from the .csv file that Jordan shared with me. It works, eventually, I will try to see if I can get that to do it off a file.